Try canHarvestBlock again. Probably will crash.

+++ jmm/mods/Diamerald/blackDiameraldpickaxe.java
@@ -71,11 +71,16 @@ public class blackDiameraldpickaxe extends ItemPickaxe {
 			for (int j = -1; j < 2; j++) {
 				for (int k = -1; k < 2; k++) {
 					if (offsetY[dir] == 0) {
+						if(this.canHarvestBlockId(Block.blocksList[par2World.getBlockId(par4 + k + Direction.offsetX[direction],
+								par5 + i,
+								par6 + j + Direction.offsetZ[direction])]))
 						par2World.destroyBlock(par4 + k
 								+ Direction.offsetX[direction], par5 + i
 								+ offsetY[dir], par6 + j
 								+ Direction.offsetZ[direction], true);
 					} else {
+						if(this.canHarvestBlock(Block.blocksList[par2World.getBlockId(par4 + k, par5 + i
+								+ offsetY[dir], par6 + j)]))
 						par2World.destroyBlock(par4 + k, par5 + i
 								+ offsetY[dir], par6 + j, true);
 					}
